auth complete, akses superadmin dan admin dah dibedakan pake middleware gate dan if, miss ketika di back home masih muncul

+++ app/Http/Controllers/AdminHomeController.php
@@ -4,12 +4,22 @@ namespace App\Http\Controllers;
 
 use App\Models\Home;
 use Illuminate\Http\Request;
+use Gate;
 
 class AdminHomeController extends Controller
 {
+    // public function __construct(){
+    //     if(!Gate::allows('isSuperAdmin')){
+    //         abort(404);
+    //     }
+    // }
 
     public function index()
     {
+        if(!Gate::allows('isSuperAdmin')){
+            abort(404);
+        }
+
         $homes = Home::all();
 
         return view('super_admin.homes.index', compact('homes'));
@@ -17,12 +27,18 @@ class AdminHomeController extends Controller
 
     public function create()
     {
+        if(!Gate::allows('isSuperAdmin')){
+            abort(404);
+        }
         return view('super_admin.homes.create');    
     }
 
     
     public function store(Request $request)
     {
+        if(!Gate::allows('isSuperAdmin')){
+            abort(404);
+        }
         $this->validateCreate($request);
 
         $home = Home::create([
@@ -41,12 +57,18 @@ class AdminHomeController extends Controller
     
     public function edit(Home $home)
     {
+        if(!Gate::allows('isSuperAdmin')){
+            abort(404);
+        }
         return view('super_admin.homes.edit', compact('home') );    
     }
 
     
     public function update(Request $request, Home $home)
     {
+        if(!Gate::allows('isSuperAdmin')){
+            abort(404);
+        }
         $this->validateUpdate($request);
 
         $home->update([
@@ -59,6 +81,9 @@ class AdminHomeController extends Controller
     
     public function destroy(Home $home)
     {
+        if(!Gate::allows('isSuperAdmin')){
+            abort(404);
+        }
         $home->delete();
         
         return redirect('admin/homes');
